# Clear side forwarding bits properly