--- a/src/util/copy/mod.rs
+++ b/src/util/copy/mod.rs
@@ -110,8 +110,15 @@ impl<VM: VMBinding> GCWorkerCopyContext<VM> {
     /// * `bytes`: The size of the object in bytes.
     /// * `semantics`: The copy semantic used for the copying.
     pub fn post_copy(&mut self, object: ObjectReference, bytes: usize, semantics: CopySemantics) {
-        // Clear forwarding bits.
-        object_forwarding::clear_forwarding_bits::<VM>(object);
+        if VM::VMObjectModel::LOCAL_FORWARDING_BITS_SPEC.is_in_header() {
+            // Clear forwarding bits if the forwarding bits are in the header.
+            object_forwarding::clear_forwarding_bits::<VM>(object);
+        } else {
+            // We ensure no stale side forwarding bits exist before tracing.
+            debug_assert!(!object_forwarding::is_forwarded_or_being_forwarded::<VM>(
+                object
+            ));
+        }
         // If we are copying objects in mature space, we would need to mark the object as mature.
         if semantics.is_mature() && self.config.constraints.needs_log_bit {
             // If the plan uses unlogged bit, we set the unlogged bit (the object is unlogged/mature)