@wks 2024-05-23
> > I'll run benchmarks using the default settings (in-header forwarding bits) to verify that there is no performance difference.
> 
> If we can use side forwarding bits for OpenJDK, it is worth measuring the performance.

I'll add a feature to OpenJDK to force the forwarding bits to be on the side.  And it shall be added to minimal-tests-openjdk.  Otherwise the CI has no way to even check if this PR is correct.