@wks 2024-05-24
There is a deadlock happening in the minimal binding test.  It is running `lusearch` in dacapo 2006.  The bug is reproducible locally, and it does not require the forwarding bits to be on the side (i.e. not related to this PR or the [OpenJDK PR](https://github.com/mmtk/mmtk-openjdk/pull/277)).  I have reported it in https://github.com/mmtk/mmtk-openjdk/issues/278.  Since the deadlock is so easy to trigger, I'll dig into that bug before continuing with this PR.