@wks 2024-05-31
The benchmark result.  Lusearch from DaCapo Chopin, 3x min heap size, 40 invocations. 3 builds:

-   build1: master
-   build2: this PR and [the mmtk-openjdk PR](https://github.com/mmtk/mmtk-openjdk/pull/277)
-   build3: this PR and [the mmtk-openjdk PR](https://github.com/mmtk/mmtk-openjdk/pull/277), using side forwarding bits

Note: There is no master branch with side forwarding bits because that'll crash.

Plots (normalised to build1): [plotty link](https://squirrel.anu.edu.au/plotty/wks/noproject/#0|lynx-2024-05-30-Thu-092115&benchmark^build^invocation^iteration^mmtk_gc&GC^time^time.other^time.stw&|10&iteration^1^4|20&1^invocation|30&1&benchmark^mmtk_gc&build;build1|41&Histogram%20(with%20CI)^build^mmtk_gc&)

![image](https://github.com/mmtk/mmtk-core/assets/370317/b9755e32-f8e5-4462-8ebd-381018fd19ee)

Plots (not normalised): [plotty link](https://squirrel.anu.edu.au/plotty/wks/noproject/#0|lynx-2024-05-30-Thu-092115&benchmark^build^invocation^iteration^mmtk_gc&GC^time^time.other^time.stw&|10&iteration^1^4|20&1^invocation|41&Histogram%20(with%20CI)^build^mmtk_gc&)

![image](https://github.com/mmtk/mmtk-core/assets/370317/dcf50607-2471-45ee-af5f-aa2eb5d208fa)

Note that the number of GC is stuck at 2047 for GenImmix.  The actual number of GCs should be much higher.

# build1 vs build2

The difference between build1 and build2 is the effect of this PR without enabling side forwarding bits.

Plots (normalised to build1): [plotty link](https://squirrel.anu.edu.au/plotty/wks/noproject/#0|lynx-2024-05-30-Thu-092115&benchmark^build^invocation^iteration^mmtk_gc&GC^time^time.other^time.stw&|10&iteration^1^4&build^2^build3|20&1^invocation|30&1&benchmark^mmtk_gc&build;build1|41&Histogram%20(with%20CI)^build^mmtk_gc&)

![image](https://github.com/mmtk/mmtk-core/assets/370317/e5089d4c-a025-4846-8f7d-5fa768f4fe0a)

Mostly unchanged.  For some plans, such as GenCopy and SemiSpace, STW time becomes higher, but other time becomes lower.  It's likely some changes in the code influenced the way it measures STW time, counting some other time into STW time.  The total time is different with non-overlapping error bars, but is within 1%.

# build2 vs build3

The difference between build2 and build3 indicates the cost of enabling side forwarding bits.

Plots (normalised to build2): [plotty link](https://squirrel.anu.edu.au/plotty/wks/noproject/#0|lynx-2024-05-30-Thu-092115&benchmark^build^invocation^iteration^mmtk_gc&GC^time^time.other^time.stw&|10&iteration^1^4&build^2^build1|20&1^invocation|30&1&benchmark^mmtk_gc&build;build2|41&Histogram%20(with%20CI)^build^mmtk_gc&)

Obviously, the number of GC is increased.  That's because the side forwarding bits occupies part of the heap space.  As a result, the effective heap size for objects becomes smaller.  The STW time increased due to increased number of GCs and the cost of clearing side forwarding bits.  The plans involving CopySpace (SemiSpace, GenCopy and GenImmix) are impacted to a greater degree.  The `time.other` is mostly unchanged, with GenImmix as an exception.  But the total time is consistently increased.

![image](https://github.com/mmtk/mmtk-core/assets/370317/84ad5cda-c25e-4455-9ca8-2e74d8115b2f)