@wks 2024-05-31
> The slowdown of using side forwarding bits is significant (build2 vs build3).

Yes.  Side forwarding bits are not free.  It is significant for CopySpace because the entire CopySpace is copied away, and the clearing involves the entire space.  But it is much lower for ImmixSpace.  For Immix and StickyImmix, the total time overhead is within 1%, and the STW time overhead is about 2%-4%.

> It would be good to measure the overhead of side forwarding bits on master -- if it cannot run with Immix or with discontiguous spaces, you can try find a configuration that runs (e.g. no compress pointer, semispace).

I am doing a 4-way comparison, comparing master-header, master-side, PR-header and PR-side, using SemiSpace and GenCopy, without compressed oops.  Let's wait for the results.