@wks 2024-05-31
I added build4: master with side forwarding bits enabled.  To avoid confusion, I'll use "master-header", "PR-header", "master-side", "PR-side" as the labels of the builds.  I re-ran lusearch, but only using SemiSpace and GenCopy which only involve CopySpace.

Plot (un-normalised): [plotty](https://squirrel.anu.edu.au/plotty/wks/noproject/#0|lynx-2024-05-31-Fri-043010&benchmark^build^invocation^iteration^mmtk_gc&GC^time^time.other^time.stw&|10&iteration^1^4|20&1^invocation|70&build^sidefwd-4way|41&Histogram%20(with%20CI)^build^mmtk_gc&)

![image](https://github.com/mmtk/mmtk-core/assets/370317/ed725941-9ed9-43c4-a24a-a7d50eb52992)

Plot (normalised to "master-header"): [plotty](https://squirrel.anu.edu.au/plotty/wks/noproject/#0|lynx-2024-05-31-Fri-043010&benchmark^build^invocation^iteration^mmtk_gc&GC^time^time.other^time.stw&|10&iteration^1^4|20&1^invocation|70&build^sidefwd-4way|30&1&benchmark^mmtk_gc&build;build1|41&Histogram%20(with%20CI)^build^mmtk_gc&)

![image](https://github.com/mmtk/mmtk-core/assets/370317/d0518a06-7a24-40e0-af65-0cae70d485fe)

The increase in the number of GC and the times due to moving the forwarding bits on the side (i.e. the ratio of `side / header` for each number) is similar for both the master branch and this PR.   (p.s. I wonder why the number of GC didn't overflow this time.)

For SemiSpace, there is a 1.7% increase in STW time and 0.5% increase in total time, and it is consistent with the result of the last 3-way comparison.