@wks 2024-06-02
Running jython locally, with SemiSpace, 76M heap size (3x min heap size w.r.t. G1 in OpenJDK), and no compressed oops:

master-header (number of GC = 1697, time.stw = 6489.81 ms, time = 19773ms)
![image](https://github.com/mmtk/mmtk-core/assets/370317/2f131084-32c3-47f6-898c-8441d755401f)

pr-side (number of GC > 2047 (counter overflowed), time.stw = 11954.42 ms, time = 24658ms)
![image](https://github.com/mmtk/mmtk-core/assets/370317/4b000540-31ac-4c86-b788-3c5e4521c644)

It looks like the 76M heap size is still too small for jython when using SemiSpace.  Individual GCs don't look very different, and the Release work packet is trivial compared to the Closure stage.  But when using side metadata, it is simply doing more GCs.  Let's zoom in to the middle of the execution.

master-header
![image](https://github.com/mmtk/mmtk-core/assets/370317/569eb50a-9261-43dd-9b07-88d06086e48c)

During some periods of execution (maybe when the program demands more memory), the program is spending almost 100% of the time doing GC, and the mutators hardly have any time to execute.  From time to time, the muators can get an extended period of execution time without triggering GC (probably when the program keeps fewer objects alive).  If the heap gets even a little bit smaller (or some portion of the heap is occupied by the side metadata), the thrashing will become significantly worse, amplifying the overhead of side forwarding bits.

Given that SemiSpace can only make use of 50% of the heap size, I doubled the heap size to 152M, and the problem became much milder.

master-header (GC = 141, time.stw = 590.08 ms, time = 13926 ms)
![image](https://github.com/mmtk/mmtk-core/assets/370317/dc1efc66-76ed-4638-bb8d-95091ff5df9a)

pr-side (GC = 149, time.stw = 735.66 ms, time = 14202 ms)
![image](https://github.com/mmtk/mmtk-core/assets/370317/b8714a90-1094-4f75-9820-40dbc47b03cc)

The STW time still increased significantly when using side metadata, but the total time is less affected.  At this heap size, it still occasionally have bursts of frequent GCs.  See the following plot:

master-header
![image](https://github.com/mmtk/mmtk-core/assets/370317/0b5b618f-eaff-49b7-92ae-c95c3330507c)

Each individual GC still takes the similar time (about 3.5ms) as 76M, but the density is much lower.  I think it is the nature of the `jython` benchmark to have short bursts of allocations, but the amount of memory it keeps alive is not very high (that's why each GC can still finish in 3.5ms regardless of heap size).  Ideally, this kind of workload could benefit from MemBalancer.

FYI, the following timelines are recorded from StickyImmix with 76M heap size:

master-header (GC = 178, time.stw = 458.02, time = 14110)
![image](https://github.com/mmtk/mmtk-core/assets/370317/48d0c4e9-30ab-43b7-a339-5dc58d23cf1e)

pr-side (GC = 186, time.stw = 423.02, time = 13823) (Note: I recorded it on my laptop and only ran once.  The speed-up may just be noise, but the *shape* of the timeline is important.)
![image](https://github.com/mmtk/mmtk-core/assets/370317/80913e42-833b-4aa7-a12a-4d3ad984799b)

zoom-in of master-header
![image](https://github.com/mmtk/mmtk-core/assets/370317/66d36c5a-c784-4433-a707-b4229058fc8f)

The allocation is still bursty, and there are some long GCs (which take up to 9.8ms).  But overall it only spent a small amount of time doing GC.